fix skeleton toolbar
