+++ frontend/javascripts/oxalis/model/reducers/annotation_reducer.js
@@ -5,7 +5,12 @@ import update from "immutability-helper";
 import type { Action } from "oxalis/model/actions/actions";
 import type { OxalisState, UserBoundingBox } from "oxalis/store";
 import { type StateShape1, updateKey, updateKey2 } from "oxalis/model/helpers/deep_update";
-import { convertServerAnnotationToFrontendAnnotation } from "oxalis/model/reducers/reducer_helpers";
+import {
+  convertServerAnnotationToFrontendAnnotation,
+  setToolReducer,
+} from "oxalis/model/reducers/reducer_helpers";
+import { AnnotationToolEnum } from "oxalis/constants";
+import { hideBrushReducer } from "oxalis/model/reducers/volumetracing_reducer_helpers";
 
 const updateTracing = (state: OxalisState, shape: StateShape1<"tracing">): OxalisState =>
   updateKey(state, "tracing", shape);
@@ -138,6 +143,24 @@ function AnnotationReducer(state: OxalisState, action: Action): OxalisState {
       return updateKey2(state, "isosurfaces", cellId, {
         isLoading: false,
       });
+    }
+
+    case "SET_TOOL": {
+      if (!state.tracing.restrictions.allowUpdate) {
+        return state;
+      }
+      return setToolReducer(state, action.tool);
+    }
+
+    case "CYCLE_TOOL": {
+      if (!state.tracing.restrictions.allowUpdate) {
+        return state;
+      }
+      const tools = Object.keys(AnnotationToolEnum);
+      const currentToolIndex = tools.indexOf(state.tracing.activeTool);
+      const newTool = tools[(currentToolIndex + 1) % tools.length];
+
+      return setToolReducer(hideBrushReducer(state), newTool);
     }
 
     default:
