+++ frontend/javascripts/oxalis/model/reducers/reducer_helpers.js
@@ -1,5 +1,6 @@
 // @flow
 import Maybe from "data.maybe";
+import update from "immutability-helper";
 
 import type {
   APIAnnotation,
@@ -11,12 +12,17 @@ import type {
   BoundingBoxObject,
   UserBoundingBox,
   UserBoundingBoxToServer,
+  OxalisState,
 } from "oxalis/store";
 import type { Boundary } from "oxalis/model/accessors/dataset_accessor";
-import type { BoundingBoxType } from "oxalis/constants";
+import type { BoundingBoxType, AnnotationTool } from "oxalis/constants";
 import { AnnotationToolEnum } from "oxalis/constants";
 import { V3 } from "libs/mjs";
 import * as Utils from "libs/utils";
+import {
+  isVolumeTool,
+  isVolumeAnnotationDisallowedForZoom,
+} from "oxalis/model/accessors/volumetracing_accessor";
 
 export function convertServerBoundingBoxToBoundingBox(
   boundingBox: ServerBoundingBox,
@@ -130,4 +136,19 @@ export function convertServerAnnotationToFrontendAnnotation(annotation: APIAnnot
     // todop: activeTool needs to be stored in annotation (on server)
     activeTool: AnnotationToolEnum.MOVE,
   };
+}
+
+export function setToolReducer(state: OxalisState, tool: AnnotationTool) {
+  if (tool === state.tracing.activeTool) {
+    return state;
+  }
+  if (isVolumeTool(tool) && isVolumeAnnotationDisallowedForZoom(tool, state)) {
+    return state;
+  }
+
+  return update(state, {
+    tracing: {
+      activeTool: { $set: tool },
+    },
+  });
 }
