+++ frontend/javascripts/oxalis/model/reducers/volumetracing_reducer_helpers.js
@@ -8,25 +8,10 @@
 
 import update from "immutability-helper";
 
-import { type ContourMode, type Vector3, type AnnotationTool } from "oxalis/constants";
+import { type ContourMode, type Vector3 } from "oxalis/constants";
 import type { OxalisState, VolumeTracing, VolumeCell } from "oxalis/store";
-import { isVolumeAnnotationDisallowedForZoom } from "oxalis/model/accessors/volumetracing_accessor";
 import { setDirectionReducer } from "oxalis/model/reducers/flycam_reducer";
-
-export function setToolReducer(state: OxalisState, volumeTracing: VolumeTracing, tool: AnnotationTool) {
-  if (tool === state.tracing.activeTool) {
-    return state;
-  }
-  if (isVolumeAnnotationDisallowedForZoom(tool, state)) {
-    return state;
-  }
-
-  return update(state, {
-    tracing: {
-      activeTool: { $set: tool },
-    },
-  });
-}
+import { isVolumeAnnotationDisallowedForZoom } from "oxalis/model/accessors/volumetracing_accessor";
 
 export function setActiveCellReducer(state: OxalisState, volumeTracing: VolumeTracing, id: number) {
   const newActiveCell = volumeTracing.cells[id];
